# api.yaml
swagger: '2.0'
info:
  description: REST API description of the ElasTest TORM Module.
  version: 1.0.0
  title: ElasTest Torm REST API
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8091'
basePath: /v2
...
paths:
...
  /sut:
    get:
      tags:
        - sut
      summary: List all SuT definitions
      operationId: getSutListUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/SutSpecification'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: SuTs not found
    post:
      tags:
        - sut
      summary: Create SuT Description
      operationId: createSuTUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: body
          description: SuT configuration
          required: true
          schema:
            $ref: '#/definitions/SutSpecification'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SutSpecification'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Invalid input
    put:
      tags:
        - sut
      summary: Updates an existing SuT Description
      operationId: modifySutUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: SuT configuration
          required: true
          schema:
            $ref: '#/definitions/SutSpecification'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SutSpecification'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Invalid input
          schema:
            $ref: '#/definitions/SutSpecification'
  '/sut/{sutId}':
    get:
      tags:
        - sut
      summary: Returns a Sut.
      description: Returns the Sut that matches the given id.
      operationId: getSutByIdUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: sutId
          in: path
          description: SuT id to return.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Sut returned successfully
          schema:
            $ref: '#/definitions/SutSpecification'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - sut
      summary: Deletes a SuT
      operationId: deleteSuTUsingDELETE
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: sutId
          in: path
          description: SuT id to delete
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: SuT deleteted successfully
          schema:
            type: integer
            format: int64
        '204':
          description: No Content
        '400':
          description: Invalid SuT ID supplied
          schema:
            type: integer
            format: int64
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: SuT not found
          schema:
            type: integer
            format: int64
  '/sut/{sutId}/exec':
    get:
      tags:
        - sut execution
      summary: List all SuT executions
      operationId: getAllSutExecBySutUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: sutId
          in: path
          description: Sut id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/SutExecution'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: SuTs not found
          schema:
            $ref: '#/definitions/SutExecution'
    post:
      tags:
        - sut execution
      summary: Deploys a SuT
      operationId: deploySuTUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: sutId
          in: path
          description: SuT id to deploy
          required: true
          type: integer
          format: int64
        - in: body
          name: deployConfig
          description: Configuration for deploy
          required: true
          schema:
            $ref: '#/definitions/DeployConfig'
      responses:
        '200':
          description: SuT has been deployed successfully
          schema:
            $ref: '#/definitions/SutExecution'
        '201':
          description: Created
        '400':
          description: Invalid SuT ID supplied
          schema:
            $ref: '#/definitions/SutExecution'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: SuT not found
          schema:
            $ref: '#/definitions/SutExecution'
  '/sut/{sutId}/exec/{sutExecId}':
    get:
      tags:
        - sut execution
      summary: Returns a SuT Execution information
      operationId: getSutExecUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: sutId
          in: path
          description: SuT id to get info
          required: true
          type: integer
          format: int64
        - name: sutExecId
          in: path
          description: SuT Execution id to get info
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SutExecution'
        '400':
          description: Invalid SuT ID supplied
          schema:
            $ref: '#/definitions/SutExecution'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: SuT not found
          schema:
            $ref: '#/definitions/SutExecution'
    delete:
      tags:
        - sut execution
      summary: Deletes a SuT execution
      operationId: deleteSuTExecUsingDELETE
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: sutExecId
          in: path
          description: SuT execution id to delete
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: SuT Execution deleteted successfully
          schema:
            type: integer
            format: int64
        '204':
          description: No Content
        '400':
          description: Invalid SuT Execution ID supplied
          schema:
            type: integer
            format: int64
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: SuT Execution not found
          schema:
            type: integer
            format: int64
  '/sut/{sutId}/exec/{sutExecId}/logs':
    get:
      tags:
        - sut execution
      summary: Returns a SuT logs
      operationId: suTLogsUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: sutId
          in: path
          description: SuT id to return logs
          required: true
          type: integer
          format: int64
        - name: sutExecId
          in: path
          description: SuT Execution id to deploy
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Log'
        '400':
          description: Invalid SuT ID supplied
          schema:
            $ref: '#/definitions/Log'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: SuT not found
          schema:
            $ref: '#/definitions/Log'
  '/sut/{sutId}/exec/{sutExecId}/monitoring':
    get:
      tags:
        - sut execution
      summary: Returns a SuT monitoring info
      operationId: suTMonitoringUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: sutId
          in: path
          description: SuT id to return monitoring information
          required: true
          type: integer
          format: int64
        - name: sutExecId
          in: path
          description: SuT Execution id to deploy
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/SuTMonitoring'
        '400':
          description: Invalid SuT ID supplied
          schema:
            $ref: '#/definitions/SuTMonitoring'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: SuT not found
          schema:
            $ref: '#/definitions/SuTMonitoring'
  '/sut/{sutId}/exec/{sutExecId}/status':
    get:
      tags:
        - sut execution
      summary: Returns a SuT status
      operationId: suTStatusUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: sutId
          in: path
          description: SuT id to return status
          required: true
          type: integer
          format: int64
        - name: sutExecId
          in: path
          description: SuT Execution id to deploy
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          schema:
            type: string
        '400':
          description: Invalid SuT ID supplied
          schema:
            type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: SuT not found
          schema:
            type: string
  '/sut/{sutId}/exec/{sutExecId}/undeploy':
    post:
      tags:
        - sut execution
      summary: Undeploys a SuT
      operationId: undeploySuTUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: sutId
          in: path
          description: SuT id to undeploy
          required: true
          type: integer
          format: int64
        - name: sutExecId
          in: path
          description: SuT Execution id to deploy
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: SuT undeployed successfully
        '201':
          description: Created
        '400':
          description: Invalid SuT ID supplied
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: SuT not found
...
definitions:
  Project:
    type: object
    required:
      - id
      - name
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      suts:
        type: array
        items:
          $ref: '#/definitions/SutSpecification'
      tJobs:
        type: array
        items:
          $ref: '#/definitions/TJob'
      tOJobs:
        type: array
        items:
          $ref: '#/definitions/TOJob'
      tjobs:
        type: array
        items:
          $ref: '#/definitions/TJob'
      tojob:
        type: array
        items:
          $ref: '#/definitions/TOJob'
      tojobs:
        type: array
        items:
          $ref: '#/definitions/TOJob'
  Log:
    type: object
    required:
      - logType
      - logUrl
    properties:
      id:
        type: integer
        format: int64
      logType:
        type: string
        enum:
          - SutLog
          - TestLog
      logUrl:
        type: string
      tJobExec:
        $ref: '#/definitions/TJobExecution'
  SutExecution:
    type: object
    properties:
      deployStatus:
        type: string
        enum:
          - deploying
          - deployed
          - undeploying
          - undeployed
          - error
      id:
        type: integer
        format: int64
        example: 12345678
      sutSpecification:
        $ref: '#/definitions/SutSpecification'
  TJobExecution:
    type: object
    required:
      - duration
      - id
      - result
    properties:
      duration:
        type: integer
        format: int64
      error:
        type: string
      id:
        type: integer
        format: int64
      logs:
        type: array
        description: URLs of logs
        items:
          $ref: '#/definitions/Log'
      result:
        type: string
        enum:
          - SUCCESS
          - FAILURE
          - IN PROGRESS
      sutExecution:
        $ref: '#/definitions/SutExecution'
      tjob:
        $ref: '#/definitions/TJob'
      tojobExecution:
        $ref: '#/definitions/TOJobExecution'
      tojobExecutionb:
        $ref: '#/definitions/TOJobExecution'
  TOJobExecution:
    type: object
    required:
      - id
      - status
      - tojob
    properties:
      id:
        type: integer
        format: int64
      status:
        type: string
        enum:
          - SUCCESS
          - FAILURE
          - IN PROGRESS
      tJobsExecutions:
        type: array
        items:
          $ref: '#/definitions/TJobExecution'
      tOJob:
        $ref: '#/definitions/TOJob'
      tjobsExecutions:
        type: array
        items:
          $ref: '#/definitions/TJobExecution'
      tojob:
        $ref: '#/definitions/TOJob'
  TOJob:
    type: object
    required:
      - name
      - project
      - toscript
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      project:
        $ref: '#/definitions/Project'
      tOScript:
        type: string
      tojobExecutions:
        type: array
        items:
          $ref: '#/definitions/TOJobExecution'
      toscript:
        type: string
        example: 'tjob1 {sut1, browser}, tjob2 {sut1, browser}, tjob3 {sut2, browser2}'
  SuTMonitoring:
    type: object
    properties:
      CPU Use:
        type: array
        items:
          type: string
      CPU cores:
        type: string
        example: '4'
      Mem Use:
        type: array
        items:
          type: string
  DeployedSut:
    type: object
    required:
      - url
    properties:
      sshKey:
        type: string
      sutSpecification:
        $ref: '#/definitions/SutSpecification'
      url:
        type: string
  SutSpecification:
    type: object
    required:
      - name
      - project
      - specification
    properties:
      deployedSut:
        type: array
        items:
          $ref: '#/definitions/DeployedSut'
      description:
        type: string
        example: This is a SuT description example
      id:
        type: integer
        format: int64
      name:
        type: string
        example: sut definition 1
      project:
        example: '12345678'
        $ref: '#/definitions/Project'
      specification:
        type: string
      sutExecution:
        type: array
        items:
          $ref: '#/definitions/SutExecution'
      tJobs:
        type: array
        items:
          $ref: '#/definitions/TJob'
      tjobs:
        type: array
        items:
          $ref: '#/definitions/TJob'
  DeployConfig:
    type: object
    properties:
      AWS CloudFormation:
        type: string
      Shell script:
        type: string
      Terraform configuration:
        type: string
  TJob:
    type: object
    required:
      - name
    properties:
      id:
        type: integer
        format: int64
      imageName:
        type: string
        example: edujgurjc/torm-test-01
      name:
        type: string
        example: testApp1
      project:
        example: '{ id:"1" }'
        $ref: '#/definitions/Project'
      sut:
        example: '{ id:"1" }'
        $ref: '#/definitions/SutSpecification'
      tjobExecs:
        type: array
        items:
          $ref: '#/definitions/TJobExecution'